Front end build comming along, App.js loading up nicely.
